Introduced BaseTelemetryProcessor
